Upgrade Doctrine Bundles, introduce new Migrations & Installer (#7543)
* Removed custom migrations and MigrationInstaller

* Fixed is installed check

* docs

* fixed stan

+++ bundles/CoreBundle/Migrations/Version20201009095924.php
@@ -3,22 +3,24 @@
 namespace Pimcore\Bundle\CoreBundle\Migrations;
 
 use Doctrine\DBAL\Schema\Schema;
-use Pimcore\Migrations\Migration\AbstractPimcoreMigration;
+use Doctrine\Migrations\AbstractMigration;
 
-class Version20201009095924 extends AbstractPimcoreMigration
+class Version20201009095924 extends AbstractMigration
 {
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE `users` DROP COLUMN `apiKey`;');
+        if($schema->getTable('users')->hasColumn('apiKey')) {
+            $this->addSql('ALTER TABLE `users` DROP COLUMN `apiKey`;');
+        }
     }
 
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema): void
     {
         $this->addSql('ALTER TABLE `users` ADD COLUMN `apiKey` varchar(255) DEFAULT NULL;');
     }
